inventory gamemenu visual only

+++ src/net/torvald/terrarum/ui/UIUtils.kt
@@ -6,6 +6,6 @@ import net.torvald.terrarum.sqr
  * Created by minjaesong on 2017-03-14.
  */
 object UIUtils {
-    fun moveQuick(start: Double, end: Double, timer: Double, duration: Double) =
+    fun moveQuick(start: Float, end: Float, timer: Float, duration: Float) =
             (start - end) * ((timer / duration) - 1).sqr() + end
 }
